add order management functionality with controllers and services

+++ src/main/java/com/kjt/demo/dto/OrderItemResponse.java
@@ -3,13 +3,15 @@ package com.kjt.demo.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class OrderItemResponse {
     private Long bookId;
     private String bookTitle;
     private Integer quantity;
-    private Double price;
-    private Double subtotal;
+    private BigDecimal price;
+    private BigDecimal subtotal;
 }
